Make windows build happy

+++ src/DigitalStateVariableFilterPlugin/Editor.cpp
@@ -3,7 +3,7 @@
 Editor::Editor(juce::AudioProcessor& process) :
   GenericAudioProcessorEditor(process)
 {
-  const int w = 850 / 1.618, h = 850;
+  const int w = int(850 / 1.618), h = 850;
 
   setSize(w, h);
 }
